# api/docs/swagger.yaml
basePath: /aiapp
definitions:
  model.AcceleratorRequirement:
    properties:
      minCount:
        description: MinCount is the smallest acceptable number of devices per node.
        example: 1
        type: integer
      minMemoryGiB:
        description: MinMemoryGiB is the smallest acceptable per-device memory. Zero
          means unset.
        example: 22
        type: number
      model:
        description: Model narrows the search to a device model. Empty means any model
          of Type.
        example: NVIDIA L4
        type: string
      type:
        description: Type is the accelerator class, matched against the spec acceleratorType.
        example: gpu
        type: string
    required:
    - minCount
    - type
    type: object
  model.AppSpec:
    properties:
      accelerator:
        $ref: '#/definitions/model.AcceleratorRequirement'
      deployTargets:
        items:
          $ref: '#/definitions/model.DeployTarget'
        type: array
      description:
        example: OpenAI compatible inference endpoint
        type: string
      id:
        example: vllm-llama31-8b
        type: string
      install:
        $ref: '#/definitions/model.InstallSpec'
      labels:
        additionalProperties:
          type: string
        type: object
      model:
        $ref: '#/definitions/model.ModelRef'
      name:
        example: vLLM Llama 3.1 8B Instruct
        type: string
      resources:
        $ref: '#/definitions/model.ResourceRequirement'
      runtime:
        allOf:
        - $ref: '#/definitions/model.Runtime'
        example: vllm
      runtimeVersion:
        example: 0.6.3
        type: string
      serving:
        $ref: '#/definitions/model.ServingSpec'
      version:
        example: 0.1.0
        type: string
    required:
    - accelerator
    - id
    - install
    - model
    - name
    - runtime
    - serving
    - version
    type: object
  model.AppSpecListResp:
    properties:
      apps:
        items:
          $ref: '#/definitions/model.AppSpec'
        type: array
    type: object
  model.ArchiveListResp:
    properties:
      records:
        items:
          $ref: '#/definitions/model.ArchiveRecord'
        type: array
    type: object
  model.ArchiveRecord:
    properties:
      answer:
        type: string
      endedAt:
        type: string
      error:
        type: string
      intent:
        type: string
      kind:
        example: intent
        type: string
      nsId:
        type: string
      runId:
        type: string
      startedAt:
        type: string
      steps:
        items:
          $ref: '#/definitions/model.ToolStep'
        type: array
      usage:
        $ref: '#/definitions/model.TokenUsage'
    type: object
...
  model.DeployTarget:
    enum:
    - vm
    - container
    type: string
    x-enum-varnames:
    - DeployTargetVM
    - DeployTargetContainer
...
  model.InstallSpec:
    properties:
      commands:
        items:
          type: string
        type: array
      osType:
        description: OSType selects the node image family. Empty uses the service
          default.
        example: ubuntu 22.04
        type: string
      timeoutMinutes:
        example: 30
        type: integer
      userName:
        example: cb-user
        type: string
    required:
    - commands
    type: object
...
  model.ModelRef:
    properties:
      format:
        description: Format is the artifact format the runtime loads.
        example: safetensors
        type: string
      id:
        description: ID is the model identifier in the AI model registry.
        example: meta-llama/Llama-3.1-8B-Instruct
        type: string
      sizeGiB:
        description: SizeGiB is the on-disk footprint, used to size the root disk.
        example: 16
        type: number
    required:
    - id
    type: object
  model.ResourceRequirement:
    properties:
      minMemoryGiB:
        example: 32
        type: integer
      minVCPU:
        example: 8
        type: integer
      rootDiskGiB:
        example: 100
        type: integer
    type: object
...
  model.Runtime:
    enum:
    - vllm
    - triton
    - tgi
    - ollama
    type: string
    x-enum-varnames:
    - RuntimeVLLM
    - RuntimeTriton
    - RuntimeTGI
    - RuntimeOllama
  model.ServingSpec:
    properties:
      healthPath:
        example: /health
        type: string
      port:
        example: 8000
        type: integer
      protocol:
        example: http
        type: string
    required:
    - port
    type: object
  model.SimpleMsg:
    properties:
      message:
        example: AI application not found
        type: string
    type: object
  model.SpecCandidate:
    properties:
      acceleratorCount:
        type: integer
      acceleratorMemoryGB:
        type: number
      acceleratorModel:
        type: string
      acceleratorType:
        type: string
      costPerHour:
        type: number
      id:
        type: string
      memoryGiB:
        type: number
      providerName:
        type: string
      regionName:
        type: string
      vCPU:
        type: integer
    type: object
  model.TokenUsage:
    properties:
      inputTokens:
        type: integer
      outputTokens:
        type: integer
    type: object
...
  model.ToolStep:
    properties:
      calledAt:
        type: string
      duration:
        type: string
      error:
        type: string
      executed:
        description: Executed is false when the dry-run guard blocked an infrastructure
          change.
        type: boolean
      grade:
        description: Grade is the tool risk grade that decided whether the call was
          executed.
        type: string
      index:
        type: integer
      input: {}
      output: {}
      tool:
        type: string
    type: object
...
info:
  contact: {}
  description: Registers AI application metadata, turns an operator instruction into
    CB-Tumblebug calls through an LLM, deploys the application onto AI semiconductor
    nodes and archives every step.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: AI-MCMP AI Application Automation Agent (Prototype)
  version: 0.1.0
paths:
  /apps:
    get:
      description: List every AI application in the catalog with its runtime, model
        and accelerator requirement
      operationId: GetAllApps
      produces:
      - application/json
      responses:
        "200":
          description: AI application list
          schema:
            $ref: '#/definitions/model.AppSpecListResp'
      summary: List registered AI applications
      tags:
      - '[App] AI Application Catalog'
    post:
      consumes:
      - application/json
      description: Register or replace an AI application metadata record. Replacing
        an existing id publishes a new version
      operationId: PostApp
      parameters:
      - description: AI application metadata
        in: body
        name: appSpec
        required: true
        schema:
          $ref: '#/definitions/model.AppSpec'
      produces:
      - application/json
      responses:
        "200":
          description: AI application registered
          schema:
            $ref: '#/definitions/model.AppSpec'
        "400":
          description: Malformed or incomplete metadata
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: Register an AI application
      tags:
      - '[App] AI Application Catalog'
  /apps/{appId}:
    delete:
      description: Remove an AI application from the catalog. Deployments already
        running are untouched
      operationId: DeleteApp
      parameters:
      - description: AI application ID
        in: path
        name: appId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: AI application deleted
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "404":
          description: AI application not found
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: Delete an AI application
      tags:
      - '[App] AI Application Catalog'
    get:
      description: Read one registered AI application, including its install commands
        and serving port
      operationId: GetApp
      parameters:
      - default: vllm-llama31-8b
        description: AI application ID
        in: path
        name: appId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: AI application metadata
          schema:
            $ref: '#/definitions/model.AppSpec'
        "404":
          description: AI application not found
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: Get an AI application
      tags:
      - '[App] AI Application Catalog'
  /apps/{appId}/specs:
    get:
      description: Find node specs across every connected cloud that satisfy the accelerator
        requirement, cheapest first
      operationId: GetAppSpecs
      parameters:
      - default: vllm-llama31-8b
        description: AI application ID
        in: path
        name: appId
        required: true
        type: string
      - default: 5
        description: How many candidates to return
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Matching node specs
          schema:
            items:
              $ref: '#/definitions/model.SpecCandidate'
            type: array
        "404":
          description: AI application not found
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "500":
          description: Spec recommendation failed
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: Recommend accelerator specs for an AI application
      tags:
      - '[App] AI Application Catalog'
  /archive:
    get:
      description: List archived deployment and control runs, newest first. A record
        holds the instruction, every tool call with its arguments and the guard verdict
      operationId: GetArchive
      parameters:
      - default: 20
        description: How many records to return
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Archived runs
          schema:
            $ref: '#/definitions/model.ArchiveListResp'
        "400":
          description: Malformed limit
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "500":
          description: Archive lookup failed
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: List archived runs
      tags:
      - '[Archive] Deploy and Control Archive'
  /archive/{runId}:
    get:
      description: Read one archived run by its identifier
      operationId: GetArchiveRun
      parameters:
      - description: Run ID
        in: path
        name: runId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Archived run
          schema:
            $ref: '#/definitions/model.ArchiveRecord'
        "404":
          description: Run not found
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "500":
          description: Archive lookup failed
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: Get one archived run
      tags:
      - '[Archive] Deploy and Control Archive'
...
swagger: "2.0"